Changed: Move matrix3d out to its own file
